+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
 
 ### fix
 
+- Create black colors from HSL color codes without NaN components.
 - Make `ColorCodeType` raw values follow the case order.
 - Recognize CSS color keyword aliases that use the "grey" spelling.